Update to FLIP v1.6 (#38)
- Flipped the ꟻ in ꟻLIP. The entire name (FLIP) should now be readable on all devices.
- Published Python version of FLIP to PyPI (URL: https://pypi.org/project/flip-evaluator/).
    - The Python version of FLIP (tool and API) is now installed by pip install flip-evaluator.
    - The distribution has been tested on Windows, Linux (Ubuntu 24.04), and OS X (≥ 10.15). Wheels are built for each (and
       various CPython versions ≥ 3.8) using cibuildwheel. Note that FLIP's output might differ slightly between the different
       operative systems. The references used for src/tests/test.py are made for Windows. While the mean tests (means
       compared up to six decimal points) pass on each mentioned operative system, not all error map pixels are identical.
    - After installation, the tool can be run directly in a shell by `flip --reference reference.{png|exr} --test test.{png|exr}`.
    - After installation, the FLIP API is available in Python by `import flip_evaluator as flip`.
- Python version is now built using scikit instead of setuptools, and uses nanobind instead of pybind11.
- Directory structure in the FLIP repository has been slightly altered to accomodate the Python version being published
   to PyPI.
- Updated Python/C++/CUDA test script.
- Various significant bugfixes.
- NOTE: Skipped version 1.5 due to PyPI-related mistake. Version 1.6 is the same as version 1.5 was supposed to be.

+++ src/cpp/tool/filename.h
@@ -170,7 +170,7 @@ namespace FLIP
             int fileNameStringLength = int(fileNameString.length());
             int totalLength = directoryStringLength + fileNameStringLength;
 
-            if (totalLength > maxLen)
+            if (size_t(totalLength) > maxLen)
             {
                 //  First shorten the directory.
                 int overflow = totalLength - int(maxLen);
@@ -418,7 +418,6 @@ namespace FLIP
                 mExtension = path.substr(1);
                 return true;
             }
-
 
             // Must be at least one slash to contain a directory.
             size_t iLastSlash = path.find_last_of("\\/");
